Only test installing keyutils if the stemcell is not xenial
Keyutils is only a dependency of cifs-utils on newer versions of
ubuntu[1]

From a previous commit, we have a test that checks if keyutils has
been installed (by checking that `keyctl -- version` succeeds). This
commit runs that check only if the stemcell we're deploying to
is *not* a xenial one.

https://bugs.launchpad.net/ubuntu/+source/cifs-utils/+bug/1772148