Support bionic and jammy (#42)
* Make smb-volume-release-job-tests use bionic-go_agent

[#181714428]

* Add test to check wheter keyutils package is being installed

[#181714428]

* Support bionic and jammy

[#181711707]

As suggested in PR #16 we are installing `keyutils` deb.

To follow a similar approach to `nfs-volume-release` we have
created the pseudo package `smb-debs` which contains the blobs.

The keyutils blobs come from http://mirrors.kernel.org/ubuntu/

* Revert "Make smb-volume-release-job-tests use bionic-go_agent"

This reverts commit 0fd01e0a674d8c005bcfd0daae9da7011f2af25a.

* Only test installing keyutils if the stemcell is not xenial

Keyutils is only a dependency of cifs-utils on newer versions of
ubuntu[1]

From a previous commit, we have a test that checks if keyutils has
been installed (by checking that `keyctl -- version` succeeds). This
commit runs that check only if the stemcell we're deploying to
is *not* a xenial one.

https://bugs.launchpad.net/ubuntu/+source/cifs-utils/+bug/1772148

Co-authored-by: Gareth Smith <sgareth@vmware.com>