Be clear about who we cf-auth as when deploying
In our smbbrokerpush errand, we auth up a CF CLI, and push some
apps. There are two options about which user we auth up as.

This commit adds echo lines so that we can see in the task output
which user we used to create the `smb` space, and push the app.

[#183259435]