fix: remove config file on stop
When the SMB driver starts, it creates a config file called
`/var/vcap/data/voldrivers/smbdriver.json` which contains the network
address of the driver amongst other things. This file should be cleaned
up when the SMB driver stops so that Diego will no longer try to
communicate with the driver.

[#185229779]