use latest available linuxfs  stack for push
[ #185663468 ]
[ #185665679 ]

when the platform upgrades stacks from one version to another, and
changes the default stack for apps,
then new apps will use the default stack
but existing apps will continue using the previously used stack ( the
old one)

We wan't to ensure that the smbbroker is always running with the latest
available stack and this commit does just that.

In case of smbbroker we push an app of the same name, thus it isn't
considered a new app but rather an existing app which is updated.
This means that for smbbroker the new stack will not be picked up
because it already has an assigned stack. Unfortunately that means that
existing deployments will not get their stack upgraded via the
previous version of the push script. It seems that the simplest solution
to this is to query the available stack available in CF and then using
the latest stack from the output. Piping the output of `cf stacks`
through grep and sort should be good enough to achieve this without
adding much complexity.

Tested with cfcli v6, v7 and v8.