fix race condition
[#185872275]

the tests are being smart and override /bin/sleep to be empty. They're
apparently doing so to NOT wait the default 15 mins of the drain script.

the problem was that the `bosh stop` got executed quicker than the fake
`rep` process was setup via bosh ssh. this adds a wait for the fake `rep`
process.