adjust for changed contents of tarball
when the autobumper for keyutils was added concourse git resource
was chosen as opposed to the initial from source compilation tarball
that was added manually in the [compile from source pr](https://github.com/cloudfoundry/smb-volume-release/pull/292)

The git resource pulls an unpacked git repo (and that repo is used to
create a tarball). This tarball does not have a single directory at it's root
but contains the same structure as the git repo root. This adjusts for
this changed structure.