add tests to ensure that the required files are copied
Since we started compiling cifs mount helpers and keyutils from
source, we need to have specific files available in system paths.

Unfortunately the mounts will break if those files are not in the
expected locations. These tests leave a breadcrump for the expected
files and their locations ( on top of actually testing that the pre-start
script does the expected copying )