add tests to ensure that the required files are copied (#367)
* add tests to ensure that the required files are copied

Since we started compiling cifs mount helpers and keyutils from
source, we need to have specific files available in system paths.

Unfortunately the mounts will break if those files are not in the
expected locations. These tests leave a breadcrump for the expected
files and their locations ( on top of actually testing that the pre-start
script does the expected copying )

* Copy context from commit message to the test

Signed-off-by: Diego Lemos <dlemos@vmware.com>

---------

Signed-off-by: Diego Lemos <dlemos@vmware.com>
Co-authored-by: Iain Findlay <iain.findlay@broadcom.com>
Co-authored-by: Gareth Smith <sgareth@vmware.com>